+++ .isort.cfg
@@ -1,4 +1,4 @@
 [tool.isort]
 profile = "black"
 [settings]
-known_third_party = aiohttp,aioredis,api,auth_config,bcrypt,core,db,db_models,debug_toolbar,decorators,django,dotenv,elasticsearch,fastapi,flasgger,flask,flask_jwt_extended,flask_sqlalchemy,group_bp,groups_bp,models,movies,multidict,orjson,password_hash,pg_to_es,psycopg2,pydantic,pytest,requests,resources,services,settings,sqlalchemy,state,test_bp,user_bp,users_bp,uvicorn,werkzeug
+known_third_party = aiohttp,aioredis,api,auth_config,bcrypt,core,db,db_models,debug_toolbar,decorators,django,dotenv,elasticsearch,fastapi,flasgger,flask,flask_jwt_extended,flask_sqlalchemy,groups_bp,models,movies,multidict,orjson,password_hash,pg_to_es,psycopg2,pydantic,pytest,requests,resources,services,settings,sqlalchemy,state,test_bp,users_bp,uvicorn,werkzeug
