+++ tests/auth_api/conftest.py
@@ -1,10 +1,8 @@
 import os
+from uuid import uuid1
 
 import pytest
 import requests
-
-from uuid import uuid1
-
 
 AUTH_API_HOST = os.getenv("AUTH_API_HOST", "flask_auth_api:5000")
 
@@ -17,20 +15,24 @@ def seven_little_guys(request):
     for _ in range(7):
         uids.append(uuid1())
     ans = requests.post(
-        f"http://{AUTH_API_HOST}/v1/user/login?login=admin&password={os.getenv('ADMIN_PASSWORD')}"
+        f"http://{AUTH_API_HOST}/v1/users/login?login=admin&password={os.getenv('ADMIN_PASSWORD')}"
     )
     assert ans.status_code == 200
     data = ans.json()
     token = data["access_token"]
     ans = requests.post(
-        f"http://{AUTH_API_HOST}/v1/group/",
-        json={"id": str(gid), "name": "Семь гномов", "description": "Тестовая группа из семи пользователей"},
-        headers={"Authorization": "Bearer " + token}
+        f"http://{AUTH_API_HOST}/v1/groups/",
+        json={
+            "id": str(gid),
+            "name": "Семь гномов",
+            "description": "Тестовая группа из семи пользователей",
+        },
+        headers={"Authorization": "Bearer " + token},
     )
     assert ans.status_code == 200
     for (i, uid) in enumerate(uids):
         ans = requests.post(
-            f"http://{AUTH_API_HOST}/v1/user/register",
+            f"http://{AUTH_API_HOST}/v1/users/register",
             json={
                 "id": str(uid),
                 "login": f"Dwarf-{i + 1}",
@@ -42,10 +44,8 @@ def seven_little_guys(request):
         )
         assert ans.status_code == 200
         ans = requests.post(
-            f"http://{AUTH_API_HOST}/v1/group/{gid}/users/",
-            json={
-                "user_id": str(uid)
-            },
+            f"http://{AUTH_API_HOST}/v1/groups/{gid}/users/",
+            json={"user_id": str(uid)},
             headers={"Authorization": "Bearer " + token},
         )
         assert ans.status_code == 200
@@ -53,7 +53,7 @@ def seven_little_guys(request):
     def teardown():
         """Удаляем тестовую группу, но входящие в нее пользователи не удаляются"""
         ans = requests.delete(
-            f"http://{AUTH_API_HOST}/v1/group/{gid}/",
+            f"http://{AUTH_API_HOST}/v1/groups/{gid}/",
             headers={"Authorization": "Bearer " + token},
         )
         assert ans.status_code == 200
