+++ tests/auth_api/docker-compose.yml
@@ -8,6 +8,8 @@ services:
     container_name: flask_auth_api
     env_file:
       - auth.env
+    ports:
+      - "5000:5000"
     volumes:
       - ../../auth_api/flask_app:/auth_api/flask_app:ro
     networks:
@@ -40,6 +42,8 @@ services:
       context: ../../auth_api/redis_auth
       dockerfile: Dockerfile
     container_name: redis_auth_test
+    ports:
+      - "6379:6379"
     volumes:
       - ../../auth_api/redis_auth/conf/etc:/usr/local/etc/redis
     networks:
