+++ tests/auth_api/src/test_users.py
@@ -1,8 +1,9 @@
 import os
+
 import pytest
 import requests
 
-AUTH_API_HOST = os.getenv('AUTH_API_HOST', 'flask_auth_api:5000')
+AUTH_API_HOST = os.getenv("AUTH_API_HOST", "flask_auth_api:5000")
 
 
 def test_user_list():
@@ -13,22 +14,26 @@ def test_user_list():
 
 
 def test_admin_login():
-    ans = requests.post(f"http://{AUTH_API_HOST}/v1/user/login?login=admin&password={os.getenv('ADMIN_PASSWORD')}")
+    ans = requests.post(
+        f"http://{AUTH_API_HOST}/v1/users/login?login=admin&password={os.getenv('ADMIN_PASSWORD')}"
+    )
     assert ans.status_code == 200
     data = ans.json()
-    assert 'access_token' in data
-    assert 'refresh_token' in data
+    assert "access_token" in data
+    assert "refresh_token" in data
 
 
 def test_admin_history():
     """Просмотр администратором истории своих действий"""
-    ans = requests.post(f"http://{AUTH_API_HOST}/v1/user/login?login=admin&password={os.getenv('ADMIN_PASSWORD')}")
+    ans = requests.post(
+        f"http://{AUTH_API_HOST}/v1/users/login?login=admin&password={os.getenv('ADMIN_PASSWORD')}"
+    )
     assert ans.status_code == 200
     data = ans.json()
-    token = data['access_token']
+    token = data["access_token"]
     ans = requests.get(
-        f"http://{AUTH_API_HOST}/v1/user/history",
-        headers={'Authorization': 'Bearer ' + token}
+        f"http://{AUTH_API_HOST}/v1/users/history",
+        headers={"Authorization": "Bearer " + token},
     )
     assert ans.status_code == 200
     data = ans.json()
@@ -40,13 +45,15 @@ def test_admin_history():
 
 def test_nobody_history():
     """Просмотр пользователем без прав истории своих действий"""
-    ans = requests.post(f"http://{AUTH_API_HOST}/v1/user/login?login=nobody&password={os.getenv('NOBODY_PASSWORD')}")
+    ans = requests.post(
+        f"http://{AUTH_API_HOST}/v1/users/login?login=nobody&password={os.getenv('NOBODY_PASSWORD')}"
+    )
     assert ans.status_code == 200
     data = ans.json()
-    token = data['access_token']
+    token = data["access_token"]
     ans = requests.get(
-        f"http://{AUTH_API_HOST}/v1/user/history",
-        headers={'Authorization': 'Bearer ' + token}
+        f"http://{AUTH_API_HOST}/v1/users/history",
+        headers={"Authorization": "Bearer " + token},
     )
     assert ans.status_code == 200
     data = ans.json()
@@ -58,13 +65,15 @@ def test_nobody_history():
 
 def test_nobody_history_paginated():
     """Просмотр пользователем без прав истории своих действий постранично"""
-    ans = requests.post(f"http://{AUTH_API_HOST}/v1/user/login?login=nobody&password={os.getenv('NOBODY_PASSWORD')}")
+    ans = requests.post(
+        f"http://{AUTH_API_HOST}/v1/users/login?login=nobody&password={os.getenv('NOBODY_PASSWORD')}"
+    )
     assert ans.status_code == 200
     data = ans.json()
-    token = data['access_token']
+    token = data["access_token"]
     ans = requests.get(
-        f"http://{AUTH_API_HOST}/v1/user/history?page_size=5&page_number=1",
-        headers={'Authorization': 'Bearer ' + token}
+        f"http://{AUTH_API_HOST}/v1/users/history?page_size=5&page_number=1",
+        headers={"Authorization": "Bearer " + token},
     )
     assert ans.status_code == 200
     data = ans.json()
